+++ django/pokemon/urls.py
@@ -4,5 +4,6 @@ from . import views
 
 urlpatterns = [
     path('', views.index),
+    path('<int:pokemon_id>', views.detail),
     path('create', views.create)
 ]