+++ django/pokemon/views.py
@@ -5,12 +5,19 @@ from rest_framework.response import Response
 
 from .models import Pokemon
 from pokedex.models import PokedexCreature
-
+from .serializers import PokemonSerializer
 
 @api_view()
 def index(request):
     pokemons = Pokemon.objects.all().values()
     return JsonResponse(list(pokemons), safe=False)
+
+
+@api_view()
+def detail(request, pokemon_id):
+    creature = Pokemon.objects.get(pk=pokemon_id)
+    serializer = PokemonSerializer(creature)
+    return Response(serializer.data)
 
 
 @api_view(['POST'])